refactor: use MergeHelmValues in helm workload instance reconciler

+++ internal/helm-workload/helm_workload_instance.go
@@ -10,7 +10,6 @@ import (
 	"helm.sh/helm/v3/pkg/action"
 	"helm.sh/helm/v3/pkg/chart/loader"
 	"helm.sh/helm/v3/pkg/cli"
-	"helm.sh/helm/v3/pkg/cli/values"
 	"helm.sh/helm/v3/pkg/getter"
 	"helm.sh/helm/v3/pkg/registry"
 	"helm.sh/helm/v3/pkg/releaseutil"
@@ -97,20 +96,19 @@ func helmWorkloadInstanceCreated(
 	}
 
 	install.ReleaseName = helmReleaseName(helmWorkloadInstance)
-	if helmWorkloadInstance.HelmReleaseNamespace != nil && *helmWorkloadInstance.HelmReleaseNamespace != "" {
-		install.Namespace = *helmWorkloadInstance.HelmReleaseNamespace
+	if helmWorkloadInstance.ReleaseNamespace != nil && *helmWorkloadInstance.ReleaseNamespace != "" {
+		install.Namespace = *helmWorkloadInstance.ReleaseNamespace
 	} else {
 		generatedNamespace := fmt.Sprintf("%s-%s", *helmWorkloadInstance.Name, util.RandomAlphaString(10))
 		install.Namespace = generatedNamespace
-		helmWorkloadInstance.HelmReleaseNamespace = &generatedNamespace
+		helmWorkloadInstance.ReleaseNamespace = &generatedNamespace
 	}
 
 	// configure version if it is supplied by the workload definition
-	if helmWorkloadDefinition.HelmChartVersion != nil && *helmWorkloadDefinition.HelmChartVersion != "" {
-		install.Version = *helmWorkloadDefinition.HelmChartVersion
+	if helmWorkloadDefinition.ChartVersion != nil && *helmWorkloadDefinition.ChartVersion != "" {
+		install.Version = *helmWorkloadDefinition.ChartVersion
 	}
 
-	install.ReleaseName = fmt.Sprintf("%s", *helmWorkloadInstance.Name)
 	install.CreateNamespace = true
 	install.DependencyUpdate = true
 	install.PostRenderer = &ThreeportPostRenderer{
@@ -140,51 +138,26 @@ func helmWorkloadInstanceCreated(
 		return 0, fmt.Errorf("failed to load helm chart: %w", err)
 	}
 
-	// write the value files to merge as needed
-	var valueFiles []string
-	if helmWorkloadDefinition.ValuesDocument != nil {
-		filePath := fmt.Sprintf("%s/%d-definition-vals.yaml", HelmValuesDir, (*helmWorkloadInstance.ID))
-		if err := os.WriteFile(
-			filePath,
-			[]byte(*helmWorkloadDefinition.ValuesDocument),
-			0644,
-		); err != nil {
-			return 0, fmt.Errorf("failed to write values file for helm workload definition values: %w", err)
-		}
-		valueFiles = append(valueFiles, filePath)
-	}
-	if helmWorkloadInstance.ValuesDocument != nil {
-		filePath := fmt.Sprintf("%s/%d-instance-vals.yaml", HelmValuesDir, (*helmWorkloadInstance.ID))
-		if err := os.WriteFile(
-			filePath,
-			[]byte(*helmWorkloadInstance.ValuesDocument),
-			0644,
-		); err != nil {
-			return 0, fmt.Errorf("failed to values file for helm workload instance values: %w", err)
-		}
-		valueFiles = append(valueFiles, filePath)
-	}
-
-	// merge the helm values
-	var helmValues map[string]interface{}
-	if len(valueFiles) > 0 {
-		vals := values.Options{ValueFiles: valueFiles}
-		helmVals, err := vals.MergeValues(nil)
-		if err != nil {
-			return 0, fmt.Errorf("failed to merge helm values: %s", err)
-		}
-		helmValues = helmVals
+	// merge helm values
+	helmValues, err := MergeHelmValuesGo(
+		util.StringPtrToString(helmWorkloadDefinition.ValuesDocument),
+		util.StringPtrToString(helmWorkloadInstance.ValuesDocument),
+	)
+	if err != nil {
+		return 0, fmt.Errorf("failed to merge helm values: %w", err)
 	}
 
 	// deploy the helm workload
 	release, err := install.Run(chart, helmValues)
 	if err != nil {
-		if uninstallErr := uninstallHelmRelease(
-			release.Name,
-			actionConf,
-			kubeClient,
-		); err != nil {
-			return 0, fmt.Errorf("failed to uninstall helm release: %w after failed to install helm chart: %w", uninstallErr, err)
+		if release != nil {
+			if uninstallErr := uninstallHelmRelease(
+				release.Name,
+				actionConf,
+				kubeClient,
+			); err != nil {
+				return 0, fmt.Errorf("failed to uninstall helm release: %w after failed to install helm chart: %w", uninstallErr, err)
+			}
 		}
 		return 0, fmt.Errorf("failed to install helm chart: %w", err)
 	}
@@ -345,8 +318,8 @@ func helmWorkloadInstanceDeleted(
 		return 0, fmt.Errorf("failed to get helm workload instances by query string: %w", err)
 	}
 	for _, hwi := range *helmWorkloadInstances {
-		if hwi.HelmReleaseNamespace != nil &&
-			*hwi.HelmReleaseNamespace == *helmWorkloadInstance.HelmReleaseNamespace &&
+		if hwi.ReleaseNamespace != nil &&
+			*hwi.ReleaseNamespace == *helmWorkloadInstance.ReleaseNamespace &&
 			*hwi.ID != *helmWorkloadInstance.ID {
 			return 0, nil
 		}
@@ -378,7 +351,7 @@ func helmWorkloadInstanceDeleted(
 	installer.DeleteNamespaces(
 		dynamicKubeClient,
 		mapper,
-		[]string{*helmWorkloadInstance.HelmReleaseNamespace},
+		[]string{*helmWorkloadInstance.ReleaseNamespace},
 	)
 
 	return 0, nil
@@ -403,15 +376,17 @@ func uninstallHelmRelease(
 	}
 
 	// remove namespace
-	gvr := schema.GroupVersionResource{Group: "", Version: "v1", Resource: "namespaces"}
-	namespaceResource := kubeClient.Resource(gvr)
-	namespace := release.Release.Namespace
-	deletePolicy := metav1.DeletePropagationForeground
-	deleteOptions := metav1.DeleteOptions{
-		PropagationPolicy: &deletePolicy,
-	}
-	if err = namespaceResource.Delete(context.TODO(), namespace, deleteOptions); err != nil {
-		return fmt.Errorf("failed to delete helm release namespace: %w", err)
+	if release != nil && release.Release != nil {
+		gvr := schema.GroupVersionResource{Group: "", Version: "v1", Resource: "namespaces"}
+		namespaceResource := kubeClient.Resource(gvr)
+		namespace := release.Release.Namespace
+		deletePolicy := metav1.DeletePropagationForeground
+		deleteOptions := metav1.DeleteOptions{
+			PropagationPolicy: &deletePolicy,
+		}
+		if err = namespaceResource.Delete(context.TODO(), namespace, deleteOptions); err != nil {
+			return fmt.Errorf("failed to delete helm release namespace: %w", err)
+		}
 	}
 
 	return nil
